updated join game radio buttons
